+++ src/main/java/com/example/demo/repositories/IGradeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.models.Grade;
 
 @Repository
-public interface IGradeRepository extends JpaRepository<Grade, Long> {
+public interface IGradeRepository extends JpaRepository<Grade, Integer> {
 
 }
